--- now.json
@@ -1,12 +1,0 @@
-{
-  "version": 2,
-  "name": "storybook",
-  "alias": "storybook-codesandbox",
-  "builds": [
-    {
-      "src": "package.json",
-      "use": "@now/static-build",
-      "config": { "distDir": "packages/common/public" }
-    }
-  ]
-}
