+++ packages/app/src/app/overmind/effects/index.ts
@@ -14,8 +14,8 @@ export { default as http } from './http';
 export { default as analytics } from './analytics';
 export { default as notificationToast } from './notificationToast';
 export { default as vscode } from './vscode';
-export { default as vercel } from './vercel';
-export { default as netlify } from './netlify';
+export { default as vercel } from './deployment/vercel';
+export { default as netlify } from './deployment/netlify';
 export { default as prettyfier } from './prettyfier';
 export { default as zip } from './zip';
 export { default as codesandboxApi } from './codesandboxApi';
